refactor(langchain-backend-aws): consolidate s3 unit tests by concern
Merge 31 small per-concern test files into the 7 topic owners
(constructor, paths, read, ls, glob, grep, download_files) to shrink
the PR's changed-files footprint while keeping every assertion.

- read: fold binary_read_mode, read_capped_object close-paths/non-bytes
  body defense, and the offset/limit base64 warning into test_read.py
- glob: fold clear_glob_cache, `**` translation, and the pattern length
  / wildcard metachar caps into test_glob.py
- grep: fold the metachar cap and malformed-response robustness into
  test_grep.py
- ls: fold malformed-listing robustness into test_ls.py
- download_files: fold concurrency, robustness, teardown signals, and
  teardown logging into test_download_files.py
- paths: fold path/key symmetry, double-encoded traversal, prefix shape
  invariant, prefix traversal at construction, require_prefix, empty
  key defense, and the prefix-violation fail-closed contract into
  test_paths.py
- constructor: fold build_client, endpoint_url / wildcard-DNS / proxies
  SSRF, proxies_config allow-list + audit logging, extra_boto_config
  validation, and extra_boto_config immutability into test_constructor.py
